requirements.txt
langchain==0.3.7
langchain-groq==0.2.0
langchain-huggingface==0.1.2
pypdf==5.2.0
chromadb==0.4.24
python-dotenv==1.0.0
sentence-transformers>=2.6.0